Add Cobrand config to include update date in email alerts
This is useful for sites where the updates aren't published on the site
so you need a bit more context in the email alerts.

Add inlude_time_in_update_alerts to default cobrand, off as default
If this is set then the date is added after the name string in the
email, passing in 'alert' as the type to dt_prettify in order to allow
the cobrand to format it accordingly.

This does require pulling out the alert confirmed time from the database
in the send_alerts code where it wasn't before.